merge branch integration_2024-07-25_331401016834 into master

See merge request: !319

+++ service/rdsmysqlv2/api_describe_parameter_template.go
@@ -212,7 +212,7 @@ type TemplateInfoForDescribeParameterTemplateOutput struct {
 
 	NeedRestart *bool `type:"boolean"`
 
-	ParameterNum *int32 `type:"int32"`
+	ParameterNum *int64 `type:"int64"`
 
 	ProjectName *string `type:"string"`
 
@@ -264,7 +264,7 @@ func (s *TemplateInfoForDescribeParameterTemplateOutput) SetNeedRestart(v bool) 
 }
 
 // SetParameterNum sets the ParameterNum field's value.
-func (s *TemplateInfoForDescribeParameterTemplateOutput) SetParameterNum(v int32) *TemplateInfoForDescribeParameterTemplateOutput {
+func (s *TemplateInfoForDescribeParameterTemplateOutput) SetParameterNum(v int64) *TemplateInfoForDescribeParameterTemplateOutput {
 	s.ParameterNum = &v
 	return s
 }
